TTS_server/requirements.txt
absl-py==2.3.1
accelerate==1.9.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==15.0.0
budoux==0.7.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cn2an==0.5.23
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
dnspython==2.7.0
editdistance==0.8.1
einops==0.8.1
einx==0.3.0
email_validator==2.2.0
exceptiongroup==1.3.0
fast-langdetect==0.3.2


fasttext-predict==0.9.2.4
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
funasr==1.0.27
future==1.0.0
g2p-en==2.1.0
g2pK==0.9.4
g2pk2==0.0.3
gast==0.6.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
inflect==7.5.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jieba_fast==0.53
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
jpype1==1.6.0
kaldiio==2.18.1
kiwisolver==1.4.8
ko-pron==1.3
konlpy==0.6.0
lazy_loader==0.4
librosa==0.10.2
lightning-utilities==0.15.1
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
mecab-python3==1.0.10
modelscope==1.10.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0

openai-whisper==20250625
OpenCC==1.1.9
orjson==3.11.1
oss2==2.19.1
packaging==25.0
pandas==2.3.1
peft==0.17.0
pillow==10.4.0
platformdirs==4.3.8
pooch==1.8.2
proces==0.1.7
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pynndescent==0.5.13
pyopenjtalk==0.4.1
pyparsing==3.2.3
pypinyin==0.55.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytorch-wpe==0.0.1
rotary-embedding-torch==0.8.9
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
robust-downloader==0.0.2


ruff==0.12.7
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.34.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
split-lang==2.1.0
starlette==0.47.2

tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
ToJyutping==3.2.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.12.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.50.0
typeguard==4.4.4
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==12.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wordsegment==1.3.1
x-transformers==2.6.1
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
